Move POI protocol in separate file.
